# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Open Agent Harness API
  version: 0.1.0
  summary: Headless distributed agent runtime API
  description: >
    Open Agent Harness is a headless agent runtime for multi-workspace,
    distributed usage. Identity, membership, and authentication are expected
    to be managed by an external service or upstream gateway. This document
    defines the HTTP API for workspace management, session messaging, run
    lifecycle, action invocation, and SSE event streaming.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /blueprints:
    get:
      tags: [Workspaces]
      summary: List workspace blueprints
      operationId: listWorkspaceBlueprints
      responses:
        '200':
          description: Workspace blueprint list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceBlueprintList'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /blueprints/upload:
    post:
      tags: [Workspaces]
      summary: Upload workspace blueprint
      operationId: uploadWorkspaceBlueprint
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - in: query
          name: overwrite
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Workspace blueprint uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceBlueprintUploadResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /blueprints/{blueprintName}:
    delete:
      tags: [Workspaces]
      summary: Delete workspace blueprint
      operationId: deleteWorkspaceBlueprint
      parameters:
        - in: path
          name: blueprintName
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '204':
          description: Workspace blueprint deleted
        default:
          $ref: '#/components/responses/ErrorResponse'
  /model-providers:
    get:
      tags: [Models]
      summary: List supported model providers
      operationId: listModelProviders
      responses:
        '200':
          description: Supported model provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProviderList'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...
    WorkspaceBlueprint:
      type: object
      required: [name]
      properties:
        name:
          type: string
    WorkspaceBlueprintList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceBlueprint'
    WorkspaceBlueprintUploadResponse:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ModelProvider:
      type: object
      required:
        - id
        - packageName
        - description
        - requiresUrl
        - useCases
      properties:
        id:
          type: string
          enum: [openai, openai-compatible]
        packageName:
          type: string
        description:
          type: string
        requiresUrl:
          type: boolean
        useCases:
          type: array
          items:
            type: string
      additionalProperties: false
    ModelProviderList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ModelProvider'
      additionalProperties: false
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          $ref: '#/components/schemas/JsonObject'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'